refactor: extract rendering delegates from ClipboardIcon (619→247 lines)
Extract three rendering delegates and a gradient paint factory from
ClipboardIcon.java to reduce its size from 619 to 247 lines:

- GradientPaintFactory: static gradient paint factory methods (pkg-private)
- ClipboardBoardRenderer: board shadow + rectangle (4 shapes, 1 composite)
- ClipboardPaperRenderer: paper fill/stroke/fold (6 shapes, 2 composites)
- ClipboardClipRenderer: clip body/ridge/outline (5 shapes, 1 composite)

All new classes are package-private, in the same package. ClipboardIcon's
public API is unchanged — no callers need modification. Rendering is
pixel-identical: method bodies are exact transplants with no logic changes.

Resolves #662
