+++ core/ide/src/main/java/org/alice/ide/clipboard/icons/GradientPaintFactory.java
@@ -1,0 +1,92 @@
+/*******************************************************************************
+ * Copyright (c) 2006, 2015, Carnegie Mellon University. All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice,
+ *    this list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ * 3. Products derived from the software may not be called "Alice", nor may
+ *    "Alice" appear in their name, without prior written permission of
+ *    Carnegie Mellon University.
+ *
+ * 4. All advertising materials mentioning features or use of this software must
+ *    display the following acknowledgement: "This product includes software
+ *    developed by Carnegie Mellon University"
+ *
+ * 5. The gallery of art assets and animations provided with this software is
+ *    contributed by Electronic Arts Inc. and may be used for personal,
+ *    non-commercial, and academic use only. Redistributions of any program
+ *    source code that utilizes The Sims 2 Assets must also retain the copyright
+ *    notice, list of conditions and the disclaimer contained in
+ *    The Alice 3.0 Art Gallery License.
+ *
+ * DISCLAIMER:
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND.
+ * ANY AND ALL EXPRESS, STATUTORY OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY,  FITNESS FOR A
+ * PARTICULAR PURPOSE, TITLE, AND NON-INFRINGEMENT ARE DISCLAIMED. IN NO EVENT
+ * SHALL THE AUTHORS, COPYRIGHT OWNERS OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
+ * INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, PUNITIVE OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING FROM OR OTHERWISE RELATING TO
+ * THE USE OF OR OTHER DEALINGS WITH THE SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ *******************************************************************************/
+package org.alice.ide.clipboard.icons;
+
+import edu.cmu.cs.dennisc.java.lang.reflect.ReflectionUtilities;
+
+import java.awt.*;
+import java.awt.geom.AffineTransform;
+import java.awt.geom.Point2D;
+import java.lang.reflect.Constructor;
+
+/**
+ * Static factory methods for creating gradient paints via reflection.
+ * Extracted from ClipboardIcon to reduce class size.
+ */
+final class GradientPaintFactory {
+
+  private GradientPaintFactory() {
+  }
+
+  static Paint new_LinearGradientPaint(Point2D start, Point2D end, float[] fractions, Color[] colors, AffineTransform gradientTransform) {
+    try {
+      Class<?> cls = Class.forName("java.awt.LinearGradientPaint");
+      Class<?> cycleMethodCls = Class.forName("java.awt.MultipleGradientPaint$CycleMethod");
+      Class<?> colorSpaceTypeCls = Class.forName("java.awt.MultipleGradientPaint$ColorSpaceType");
+      final Object NO_CYCLE = cycleMethodCls.getField("NO_CYCLE").get(null);
+      final Object SRGB = colorSpaceTypeCls.getField("SRGB").get(null);
+      Constructor<?> cnstrctr = ReflectionUtilities.getConstructor(cls, Point2D.class, Point2D.class, float[].class, Color[].class, cycleMethodCls, colorSpaceTypeCls, AffineTransform.class);
+      return (Paint) cnstrctr.newInstance(start, end, fractions, colors, NO_CYCLE, SRGB, gradientTransform);
+    } catch (Throwable t) {
+      //t.printStackTrace();
+      return colors[0];
+      //    return new java.awt.LinearGradientPaint( start, end, fractions, colors, java.awt.MultipleGradientPaint.CycleMethod.NO_CYCLE, java.awt.MultipleGradientPaint.ColorSpaceType.SRGB, gradientTransform );
+    }
+  }
+
+  static Paint new_RadialGradientPaint(Point2D center, float radius, Point2D focus, float[] fractions, Color[] colors, AffineTransform gradientTransform) {
+    try {
+      Class<?> cls = Class.forName("java.awt.RadialGradientPaint");
+      Class<?> cycleMethodCls = Class.forName("java.awt.MultipleGradientPaint$CycleMethod");
+      Class<?> colorSpaceTypeCls = Class.forName("java.awt.MultipleGradientPaint$ColorSpaceType");
+      final Object NO_CYCLE = cycleMethodCls.getField("NO_CYCLE").get(null);
+      final Object SRGB = colorSpaceTypeCls.getField("SRGB").get(null);
+      Constructor<?> cnstrctr = ReflectionUtilities.getConstructor(cls, Point2D.class, Float.TYPE, Point2D.class, float[].class, Color[].class, cycleMethodCls, colorSpaceTypeCls, AffineTransform.class);
+      return (Paint) cnstrctr.newInstance(center, radius, focus, fractions, colors, NO_CYCLE, SRGB, gradientTransform);
+    } catch (Throwable t) {
+      //t.printStackTrace();
+      return colors[0];
+      //return new java.awt.RadialGradientPaint( center, radius, focus, fractions, colors, java.awt.MultipleGradientPaint.CycleMethod.NO_CYCLE, java.awt.MultipleGradientPaint.ColorSpaceType.SRGB, gradientTransform );
+    }
+  }
+}
